Upgrade Specs2 (2.3.1 -> 2.3.6)
Hard to keep up with this guy. Also had to upgrade from shapeless_2.10.2 to shapeless_2.10.3 in order to get this to work but really that makes sense becuase I am using of version of Scala very similar to 2.10.3 and shapeless was only at that version before because I had some kind of issue before that probably had something to do with the Specs2 version I was using.